Give cmux a read timeout and use reuseport
We need a timeout on cmux so that we don't spike the number of go
routines waiting on data to be read from the socket.  We also shift to
using reuseport so that accept(2) only returns when there's real data on
the socket.  This reduces the need to wait on data to be read as well,
making the above timeout needed less.

Using github.com/valyala/fasthttp/reuseport may remove the need for cmux
and I'll test that later.  We may be able to return to maybe_tls
instead, but there's issues with doing reads inside of Accept and if
that blocks, you're going to lock the application.